Add files via upload
This repository compiles relevant literature for research on PLGA (poly(lactic-co-glycolic acid)), including a total of 33 articles. These publications cover the synthesis, characterization, applications, and research progress of PLGA in the biomedical field. Each article in the repository is assigned a unique index number, consistent with the numbering in the supporting information, to facilitate easy access and citation for readers.